use test name from context to provide a unique Finch :name, extract all Telemetry tests to a new test file
