+++ test/finch_test.exs
@@ -1,5 +1,5 @@
 defmodule FinchTest do
-  use ExUnit.Case, async: true
+  use FinchCase, async: true
   doctest Finch
 
   import ExUnit.CaptureIO
@@ -7,21 +7,15 @@ defmodule FinchTest do
   alias Finch.Response
   alias Finch.MockSocketServer
 
-  setup do
-    {:ok, bypass: Bypass.open()}
-  end
-
   describe "start_link/1" do
     test "raises if :name is not provided" do
       assert_raise(ArgumentError, ~r/must supply a name/, fn -> Finch.start_link([]) end)
     end
 
-    test "max_idle_time is deprecated" do
+    test "max_idle_time is deprecated", %{finch_name: finch_name} do
       msg =
         capture_io(:stderr, fn ->
-          start_supervised!(
-            {Finch, name: finch_name(), pools: %{default: [max_idle_time: 1_000]}}
-          )
+          start_supervised!({Finch, name: finch_name, pools: %{default: [max_idle_time: 1_000]}})
         end)
 
       assert String.contains?(
@@ -32,34 +26,35 @@ defmodule FinchTest do
   end
 
   describe "pool configuration" do
-    test "unconfigured", %{bypass: bypass} do
-      start_supervised!({Finch, name: finch_name()})
+    test "unconfigured", %{bypass: bypass, finch_name: finch_name} do
+      start_supervised!({Finch, name: finch_name})
       expect_any(bypass)
 
-      {:ok, %Response{}} = Finch.build(:get, endpoint(bypass)) |> Finch.request(finch_name())
-      assert [_pool] = get_pools(finch_name(), shp(bypass))
+      {:ok, %Response{}} = Finch.build(:get, endpoint(bypass)) |> Finch.request(finch_name)
+      assert [_pool] = get_pools(finch_name, shp(bypass))
 
-      {:ok, %Response{}} = Finch.build(:get, endpoint(bypass)) |> Finch.request(finch_name())
+      {:ok, %Response{}} = Finch.build(:get, endpoint(bypass)) |> Finch.request(finch_name)
     end
 
-    test "default can be configured", %{bypass: bypass} do
+    test "default can be configured", %{bypass: bypass, finch_name: finch_name} do
       {:ok, _} =
         Finch.start_link(
-          name: finch_name(),
+          name: finch_name,
           pools: %{default: [count: 5, size: 5]}
         )
 
       expect_any(bypass)
 
-      {:ok, %Response{}} = Finch.build("GET", endpoint(bypass)) |> Finch.request(finch_name())
-      pools = get_pools(finch_name(), shp(bypass))
+      {:ok, %Response{}} = Finch.build("GET", endpoint(bypass)) |> Finch.request(finch_name)
+      pools = get_pools(finch_name, shp(bypass))
       assert length(pools) == 5
     end
 
-    test "TLS options will be dropped from default if it connects to http", %{bypass: bypass} do
+    test "TLS options will be dropped from default if it connects to http",
+         %{bypass: bypass, finch_name: finch_name} do
       {:ok, _} =
         Finch.start_link(
-          name: finch_name(),
+          name: finch_name,
           pools: %{
             default: [
               count: 5,
@@ -73,32 +68,32 @@ defmodule FinchTest do
 
       # you will get badarg error if the verify option is applied to the connection.
       assert {:ok, %Response{}} =
-               Finch.build("GET", endpoint(bypass)) |> Finch.request(finch_name())
+               Finch.build("GET", endpoint(bypass)) |> Finch.request(finch_name)
     end
 
-    test "raises when invalid configuration is provided" do
+    test "raises when invalid configuration is provided", %{finch_name: finch_name} do
       assert_raise(
         NimbleOptions.ValidationError,
         ~r/expected :count to be a positive integer/,
         fn ->
-          Finch.start_link(name: finch_name(), pools: %{default: [count: :dog]})
+          Finch.start_link(name: finch_name, pools: %{default: [count: :dog]})
         end
       )
 
       assert_raise(ArgumentError, ~r/invalid destination/, fn ->
-        Finch.start_link(name: finch_name(), pools: %{invalid: [count: 5, size: 5]})
+        Finch.start_link(name: finch_name, pools: %{invalid: [count: 5, size: 5]})
       end)
     end
 
     test "pools are started based on only the {scheme, host, port} of the URLs",
-         %{bypass: bypass} do
+         %{bypass: bypass, finch_name: finch_name} do
       other_bypass = Bypass.open()
       default_bypass = Bypass.open()
       unix_socket = {:local, "/my/unix/socket"}
 
       start_supervised!(
         {Finch,
-         name: finch_name(),
+         name: finch_name,
          pools: %{
            endpoint(bypass, "/some-path") => [count: 5, size: 5],
            endpoint(other_bypass, "/some-other-path") => [count: 10, size: 10],
@@ -107,19 +102,19 @@ defmodule FinchTest do
          }}
       )
 
-      assert get_pools(finch_name(), shp(bypass)) |> length() == 5
-      assert get_pools(finch_name(), shp(other_bypass)) |> length() == 10
-      assert get_pools(finch_name(), shp({:http, unix_socket})) |> length() == 5
-      assert get_pools(finch_name(), shp({:https, unix_socket})) |> length() == 10
+      assert get_pools(finch_name, shp(bypass)) |> length() == 5
+      assert get_pools(finch_name, shp(other_bypass)) |> length() == 10
+      assert get_pools(finch_name, shp({:http, unix_socket})) |> length() == 5
+      assert get_pools(finch_name, shp({:https, unix_socket})) |> length() == 10
 
       # no pool has been started for this unconfigured shp
-      assert get_pools(finch_name(), shp(default_bypass)) |> length() == 0
+      assert get_pools(finch_name, shp(default_bypass)) |> length() == 0
     end
 
-    test "pools with an invalid URL cannot be started" do
+    test "pools with an invalid URL cannot be started", %{finch_name: finch_name} do
       assert_raise(ArgumentError, ~r/scheme is required for url: example.com/, fn ->
         Finch.start_link(
-          name: finch_name(),
+          name: finch_name,
           pools: %{
             "example.com" => [count: 5, size: 5]
           }
@@ -128,7 +123,7 @@ defmodule FinchTest do
 
       assert_raise(ArgumentError, ~r/scheme is required for url: example/, fn ->
         Finch.start_link(
-          name: finch_name(),
+          name: finch_name,
           pools: %{
             "example" => [count: 5, size: 5]
           }
@@ -137,7 +132,7 @@ defmodule FinchTest do
 
       assert_raise(ArgumentError, ~r/scheme is required for url: :443/, fn ->
         Finch.start_link(
-          name: finch_name(),
+          name: finch_name,
           pools: %{
             ":443" => [count: 5, size: 5]
           }
@@ -145,12 +140,14 @@ defmodule FinchTest do
       end)
     end
 
-    test "impossible to accidentally start multiple pools when they are dynamically started", %{
-      bypass: bypass
-    } do
+    test "impossible to accidentally start multiple pools when they are dynamically started",
+         %{
+           bypass: bypass,
+           finch_name: finch_name
+         } do
       start_supervised!(
         {Finch,
-         name: finch_name(),
+         name: finch_name,
          pools: %{
            default: [count: 5, size: 5]
          }}
@@ -160,12 +157,12 @@ defmodule FinchTest do
 
       Task.async_stream(
         1..50,
-        fn _ -> Finch.build(:get, endpoint(bypass)) |> Finch.request(finch_name()) end,
+        fn _ -> Finch.build(:get, endpoint(bypass)) |> Finch.request(finch_name) end,
         max_concurrency: 50
       )
       |> Stream.run()
 
-      assert get_pools(finch_name(), shp(bypass)) |> length() == 5
+      assert get_pools(finch_name, shp(bypass)) |> length() == 5
     end
   end
 
@@ -184,8 +181,8 @@ defmodule FinchTest do
   end
 
   describe "request/3" do
-    test "successful get request, with query string", %{bypass: bypass} do
-      start_supervised!({Finch, name: finch_name()})
+    test "successful get request, with query string", %{bypass: bypass, finch_name: finch_name} do
+      start_supervised!({Finch, name: finch_name})
       query_string = "query=value"
 
       Bypass.expect_once(bypass, "GET", "/", fn conn ->
@@ -195,11 +192,14 @@ defmodule FinchTest do
 
       assert {:ok, %{status: 200}} =
                Finch.build(:get, endpoint(bypass, "?" <> query_string))
-               |> Finch.request(finch_name())
+               |> Finch.request(finch_name)
     end
 
-    test "successful post request, with body and query string", %{bypass: bypass} do
-      start_supervised!({Finch, name: finch_name()})
+    test "successful post request, with body and query string", %{
+      bypass: bypass,
+      finch_name: finch_name
+    } do
+      start_supervised!({Finch, name: finch_name})
 
       req_body = "{\"response\":\"please\"}"
       response_body = "{\"right\":\"here\"}"
@@ -223,15 +223,17 @@ defmodule FinchTest do
                  [{header_key, header_val}],
                  req_body
                )
-               |> Finch.request(finch_name())
+               |> Finch.request(finch_name)
 
       assert {"content-type", "application/json"} in headers
     end
 
-    test "successful post HTTP/1 streaming request, with streaming body and query string", %{
-      bypass: bypass
-    } do
-      start_supervised!({Finch, name: finch_name()})
+    test "successful post HTTP/1 streaming request, with streaming body and query string",
+         %{
+           bypass: bypass,
+           finch_name: finch_name
+         } do
+      start_supervised!({Finch, name: finch_name})
 
       req_stream = Stream.map(1..10_000, fn _ -> "please" end)
       req_body = req_stream |> Enum.join("")
@@ -256,17 +258,19 @@ defmodule FinchTest do
                  [{header_key, header_val}],
                  {:stream, req_stream}
                )
-               |> Finch.request(finch_name())
+               |> Finch.request(finch_name)
 
       assert {"content-type", "application/json"} in headers
     end
 
-    test "successful post HTTP/2 streaming request, with streaming body and query string", %{
-      bypass: bypass
-    } do
+    test "successful post HTTP/2 streaming request, with streaming body and query string",
+         %{
+           bypass: bypass,
+           finch_name: finch_name
+         } do
       start_supervised!(
         {Finch,
-         name: finch_name(),
+         name: finch_name,
          pools: %{
            default: [
              protocol: :http2,
@@ -305,17 +309,19 @@ defmodule FinchTest do
                  [{header_key, header_val}],
                  {:stream, req_stream}
                )
-               |> Finch.request(finch_name())
+               |> Finch.request(finch_name)
 
       assert {header_key, header_val} in headers
     end
 
-    test "successful post HTTP/2 with a large binary body", %{
-      bypass: bypass
-    } do
+    test "successful post HTTP/2 with a large binary body",
+         %{
+           bypass: bypass,
+           finch_name: finch_name
+         } do
       start_supervised!(
         {Finch,
-         name: finch_name(),
+         name: finch_name,
          pools: %{
            default: [
              protocol: :http2,
@@ -353,17 +359,19 @@ defmodule FinchTest do
                  [{header_key, header_val}],
                  req_body
                )
-               |> Finch.request(finch_name())
+               |> Finch.request(finch_name)
 
       assert {header_key, header_val} in headers
     end
 
-    test "successful post HTTP/2 with a large iolist body", %{
-      bypass: bypass
-    } do
+    test "successful post HTTP/2 with a large iolist body",
+         %{
+           bypass: bypass,
+           finch_name: finch_name
+         } do
       start_supervised!(
         {Finch,
-         name: finch_name(),
+         name: finch_name,
          pools: %{
            default: [
              protocol: :http2,
@@ -401,13 +409,14 @@ defmodule FinchTest do
                  [{header_key, header_val}],
                  req_body
                )
-               |> Finch.request(finch_name())
+               |> Finch.request(finch_name)
 
       assert {header_key, header_val} in headers
     end
 
-    test "successful get request, with query string, when given a %URI{}", %{bypass: bypass} do
-      start_supervised!({Finch, name: finch_name()})
+    test "successful get request, with query string, when given a %URI{}",
+         %{bypass: bypass, finch_name: finch_name} do
+      start_supervised!({Finch, name: finch_name})
       query_string = "query=value"
       uri = URI.parse(endpoint(bypass, "?" <> query_string))
 
@@ -416,26 +425,26 @@ defmodule FinchTest do
         Plug.Conn.send_resp(conn, 200, "OK")
       end)
 
-      assert {:ok, %{status: 200}} = Finch.build(:get, uri) |> Finch.request(finch_name())
+      assert {:ok, %{status: 200}} = Finch.build(:get, uri) |> Finch.request(finch_name)
     end
 
-    test "successful get request to a unix socket" do
+    test "successful get request to a unix socket", %{finch_name: finch_name} do
       {:ok, {:local, socket_path}} = MockSocketServer.start()
 
-      start_supervised!({Finch, name: finch_name()})
+      start_supervised!({Finch, name: finch_name})
 
       assert {:ok, %Response{status: 200}} =
                Finch.build(:get, "http://localhost/", [], nil, unix_socket: socket_path)
-               |> Finch.request(finch_name())
+               |> Finch.request(finch_name)
     end
 
     @tag :capture_log
-    test "successful get request to a unix socket with tls" do
+    test "successful get request to a unix socket with tls", %{finch_name: finch_name} do
       {:ok, socket_address = {:local, socket_path}} = MockSocketServer.start(ssl?: true)
 
       start_supervised!(
         {Finch,
-         name: finch_name(),
+         name: finch_name,
          pools: %{
            {:https, socket_address} => [conn_opts: [transport_opts: [verify: :verify_none]]]
          }}
@@ -443,11 +452,11 @@ defmodule FinchTest do
 
       assert {:ok, %Response{status: 200}} =
                Finch.build(:get, "https://localhost/", [], nil, unix_socket: socket_path)
-               |> Finch.request(finch_name())
+               |> Finch.request(finch_name)
     end
 
-    test "properly handles connection: close", %{bypass: bypass} do
-      start_supervised!({Finch, name: finch_name()})
+    test "properly handles connection: close", %{bypass: bypass, finch_name: finch_name} do
+      start_supervised!({Finch, name: finch_name})
 
       Bypass.expect(bypass, fn conn ->
         conn
@@ -463,12 +472,12 @@ defmodule FinchTest do
         )
 
       for _ <- 1..10 do
-        assert {:ok, %Response{status: 200, body: "OK"}} = Finch.request(request, finch_name())
+        assert {:ok, %Response{status: 200, body: "OK"}} = Finch.request(request, finch_name)
       end
     end
 
-    test "returns error when request times out", %{bypass: bypass} do
-      start_supervised!({Finch, name: finch_name()})
+    test "returns error when request times out", %{bypass: bypass, finch_name: finch_name} do
+      start_supervised!({Finch, name: finch_name})
 
       timeout = 100
 
@@ -479,38 +488,38 @@ defmodule FinchTest do
 
       assert {:error, %{reason: :timeout}} =
                Finch.build(:get, endpoint(bypass))
-               |> Finch.request(finch_name(), receive_timeout: timeout)
+               |> Finch.request(finch_name, receive_timeout: timeout)
 
       assert {:ok, %Response{}} =
                Finch.build(:get, endpoint(bypass))
-               |> Finch.request(finch_name(), receive_timeout: timeout * 2)
+               |> Finch.request(finch_name, receive_timeout: timeout * 2)
     end
 
-    test "returns error when requesting bad address" do
-      start_supervised!({Finch, name: finch_name()})
+    test "returns error when requesting bad address", %{finch_name: finch_name} do
+      start_supervised!({Finch, name: finch_name})
 
       assert {:error, %{reason: :nxdomain}} =
-               Finch.build(:get, "http://idontexist.wat") |> Finch.request(finch_name())
+               Finch.build(:get, "http://idontexist.wat") |> Finch.request(finch_name)
     end
 
-    test "worker exits when pool times out", %{bypass: bypass} do
-      start_supervised!({Finch, name: finch_name()})
+    test "worker exits when pool times out", %{bypass: bypass, finch_name: finch_name} do
+      start_supervised!({Finch, name: finch_name})
       expect_any(bypass)
 
-      {:ok, %Response{}} = Finch.build(:get, endpoint(bypass)) |> Finch.request(finch_name())
+      {:ok, %Response{}} = Finch.build(:get, endpoint(bypass)) |> Finch.request(finch_name)
 
-      :sys.suspend(finch_name())
+      :sys.suspend(finch_name)
 
       assert {:timeout, _} =
                catch_exit(
                  Finch.build(:get, endpoint(bypass))
-                 |> Finch.request(finch_name(), pool_timeout: 0)
+                 |> Finch.request(finch_name, pool_timeout: 0)
                )
 
-      :sys.resume(finch_name())
+      :sys.resume(finch_name)
 
       assert {:ok, %Response{}} =
-               Finch.build(:get, endpoint(bypass)) |> Finch.request(finch_name(), pool_timeout: 1)
+               Finch.build(:get, endpoint(bypass)) |> Finch.request(finch_name, pool_timeout: 1)
     end
   end
 
@@ -525,474 +534,19 @@ defmodule FinchTest do
       stop_supervised(Finch)
     end
 
-    test "caller is unable to override mode", %{bypass: bypass} do
+    test "caller is unable to override mode", %{bypass: bypass, finch_name: finch_name} do
       start_supervised!(
-        {Finch, name: finch_name(), pools: %{default: [conn_opts: [mode: :active]]}}
+        {Finch, name: finch_name, pools: %{default: [conn_opts: [mode: :active]]}}
       )
 
       expect_any(bypass)
-      assert {:ok, _} = Finch.build(:get, endpoint(bypass)) |> Finch.request(finch_name())
-    end
-  end
-
-  describe "telemetry" do
-    setup %{bypass: bypass} do
-      Bypass.expect_once(bypass, "GET", "/", fn conn ->
-        Plug.Conn.send_resp(conn, 200, "OK")
-      end)
-
-      client = finch_name()
-      start_supervised!({Finch, name: client})
-
-      {:ok, client: client}
-    end
-
-    test "reports request spans", %{bypass: bypass, client: client} do
-      {test_name, _arity} = __ENV__.function
-
-      parent = self()
-      ref = make_ref()
-
-      handler = fn event, measurements, meta, _config ->
-        case event do
-          [:finch, :request, :start] ->
-            assert is_integer(measurements.system_time)
-            assert meta.name == :finch_name
-            assert %Finch.Request{} = meta.request
-
-            send(parent, {ref, :start})
-
-          [:finch, :request, :stop] ->
-            assert is_integer(measurements.duration)
-            assert meta.name == :finch_name
-            assert %Finch.Request{} = meta.request
-
-            assert {:ok, %Finch.Response{body: "OK", status: 200}} = meta.result
-
-            send(parent, {ref, :stop})
-
-          [:finch, :request, :exception] ->
-            assert is_integer(measurements.duration)
-            assert meta.name == :finch_name
-            assert %Finch.Request{} = meta.request
-            assert meta.kind == :exit
-            assert {:timeout, _} = meta.reason
-            assert meta.stacktrace != nil
-
-            send(parent, {ref, :exception})
-
-          _ ->
-            flunk("Unknown event")
-        end
-      end
-
-      :telemetry.attach_many(
-        to_string(test_name),
-        [
-          [:finch, :request, :start],
-          [:finch, :request, :stop],
-          [:finch, :request, :exception]
-        ],
-        handler,
-        nil
-      )
-
-      assert {:ok, %{status: 200}} = Finch.build(:get, endpoint(bypass)) |> Finch.request(client)
-      assert_receive {^ref, :start}
-      assert_receive {^ref, :stop}
-
-      Bypass.down(bypass)
-
-      try do
-        Finch.build(:get, endpoint(bypass)) |> Finch.request(client, pool_timeout: 0)
-      catch
-        :exit, reason ->
-          assert {:timeout, _} = reason
-      end
-
-      assert_receive {^ref, :start}
-      assert_receive {^ref, :exception}
-
-      :telemetry.detach(to_string(test_name))
-    end
-
-    test "reports queue spans", %{bypass: bypass, client: client} do
-      {test_name, _arity} = __ENV__.function
-
-      parent = self()
-      ref = make_ref()
-
-      handler = fn event, measurements, meta, _config ->
-        case event do
-          [:finch, :queue, :start] ->
-            assert is_integer(measurements.system_time)
-            assert is_pid(meta.pool)
-            assert %Finch.Request{} = meta.request
-            send(parent, {ref, :start})
-
-          [:finch, :queue, :stop] ->
-            assert is_integer(measurements.duration)
-            assert is_integer(measurements.idle_time)
-            assert is_pid(meta.pool)
-            assert %Finch.Request{} = meta.request
-            send(parent, {ref, :stop})
-
-          [:finch, :queue, :exception] ->
-            assert is_integer(measurements.duration)
-            assert is_pid(meta.pool)
-            assert meta.kind == :exit
-            assert {:timeout, _} = meta.reason
-            assert meta.stacktrace != nil
-            assert %Finch.Request{} = meta.request
-            send(parent, {ref, :exception})
-
-          _ ->
-            flunk("Unknown event")
-        end
-      end
-
-      :telemetry.attach_many(
-        to_string(test_name),
-        [
-          [:finch, :queue, :start],
-          [:finch, :queue, :stop],
-          [:finch, :queue, :exception]
-        ],
-        handler,
-        nil
-      )
-
-      assert {:ok, %{status: 200}} = Finch.build(:get, endpoint(bypass)) |> Finch.request(client)
-      assert_receive {^ref, :start}
-      assert_receive {^ref, :stop}
-
-      Bypass.down(bypass)
-
-      try do
-        Finch.build(:get, endpoint(bypass)) |> Finch.request(client, pool_timeout: 0)
-      catch
-        :exit, reason ->
-          assert {:timeout, _} = reason
-      end
-
-      assert_receive {^ref, :start}
-      assert_receive {^ref, :exception}
-
-      :telemetry.detach(to_string(test_name))
-    end
-
-    test "reports connection spans", %{bypass: bypass, client: client} do
-      {test_name, _arity} = __ENV__.function
-      parent = self()
-      ref = make_ref()
-
-      handler = fn event, measurements, meta, _config ->
-        case event do
-          [:finch, :connect, :start] ->
-            assert is_integer(measurements.system_time)
-            assert is_atom(meta.scheme)
-            assert is_integer(meta.port)
-            assert is_binary(meta.host)
-            send(parent, {ref, :start})
-
-          [:finch, :connect, :stop] ->
-            assert is_integer(measurements.duration)
-            assert is_atom(meta.scheme)
-            assert is_integer(meta.port)
-            assert is_binary(meta.host)
-            send(parent, {ref, :stop})
-
-          _ ->
-            flunk("Unknown event")
-        end
-      end
-
-      :telemetry.attach_many(
-        to_string(test_name),
-        [
-          [:finch, :connect, :start],
-          [:finch, :connect, :stop]
-        ],
-        handler,
-        nil
-      )
-
-      assert {:ok, %{status: 200}} = Finch.build(:get, endpoint(bypass)) |> Finch.request(client)
-      assert_receive {^ref, :start}
-      assert_receive {^ref, :stop}
-
-      :telemetry.detach(to_string(test_name))
-    end
-
-    test "reports send spans", %{bypass: bypass, client: client} do
-      {test_name, _arity} = __ENV__.function
-
-      parent = self()
-      ref = make_ref()
-
-      handler = fn event, measurements, meta, _config ->
-        case event do
-          [:finch, :send, :start] ->
-            assert is_integer(measurements.system_time)
-            assert is_integer(measurements.idle_time)
-            assert %Finch.Request{} = meta.request
-            send(parent, {ref, :start})
-
-          [:finch, :send, :stop] ->
-            assert is_integer(measurements.duration)
-            assert is_integer(measurements.idle_time)
-            assert %Finch.Request{} = meta.request
-            send(parent, {ref, :stop})
-
-          _ ->
-            flunk("Unknown event")
-        end
-      end
-
-      :telemetry.attach_many(
-        to_string(test_name),
-        [
-          [:finch, :send, :start],
-          [:finch, :send, :stop],
-          [:finch, :send, :exception]
-        ],
-        handler,
-        nil
-      )
-
-      assert {:ok, %{status: 200}} = Finch.build(:get, endpoint(bypass)) |> Finch.request(client)
-      assert_receive {^ref, :start}
-      assert_receive {^ref, :stop}
-
-      :telemetry.detach(to_string(test_name))
-    end
-
-    test "reports recv spans", %{bypass: bypass, client: client} do
-      {test_name, _arity} = __ENV__.function
-      parent = self()
-      ref = make_ref()
-
-      handler = fn event, measurements, meta, _config ->
-        case event do
-          [:finch, :recv, :start] ->
-            assert is_integer(measurements.system_time)
-            assert is_integer(measurements.idle_time)
-            assert %Finch.Request{} = meta.request
-            send(parent, {ref, :start})
-
-          [:finch, :recv, :stop] ->
-            assert is_integer(measurements.duration)
-            assert is_integer(measurements.idle_time)
-            assert %Finch.Request{} = meta.request
-            assert is_integer(meta.status)
-            assert is_list(meta.headers)
-            send(parent, {ref, :stop})
-
-          _ ->
-            flunk("Unknown event")
-        end
-      end
-
-      :telemetry.attach_many(
-        to_string(test_name),
-        [
-          [:finch, :recv, :start],
-          [:finch, :recv, :stop]
-        ],
-        handler,
-        nil
-      )
-
-      assert {:ok, %{status: 200}} = Finch.build(:get, endpoint(bypass)) |> Finch.request(client)
-      assert_receive {^ref, :start}
-      assert_receive {^ref, :stop}
-
-      :telemetry.detach(to_string(test_name))
-    end
-  end
-
-  describe "telemetry events" do
-    setup %{bypass: bypass} do
-      Bypass.expect(bypass, "GET", "/", fn conn ->
-        Plug.Conn.send_resp(conn, 200, "OK")
-      end)
-
-      client = finch_name()
-      start_supervised!({Finch, name: client, pools: %{default: [conn_max_idle_time: 10]}})
-
-      {:ok, client: client}
-    end
-
-    test "reports request and response headers", %{bypass: bypass, client: client} do
-      {test_name, _arity} = __ENV__.function
-      self = self()
-
-      :telemetry.attach_many(
-        to_string(test_name),
-        [[:finch, :send, :start], [:finch, :recv, :stop]],
-        fn name, _, metadata, _ -> send(self, {:telemetry_event, name, metadata}) end,
-        nil
-      )
-
-      Bypass.expect(bypass, "GET", "/", fn conn ->
-        conn
-        |> Plug.Conn.put_resp_header("x-foo-response", "bar-response")
-        |> Plug.Conn.send_resp(200, "OK")
-      end)
-
-      request = Finch.build(:get, endpoint(bypass), [{"x-foo-request", "bar-request"}])
-      assert {:ok, %{status: 200}} = Finch.request(request, client)
-
-      assert_receive {:telemetry_event, [:finch, :send, :start],
-                      %{request: %{headers: [{"x-foo-request", "bar-request"}]}}}
-
-      assert_receive {:telemetry_event, [:finch, :recv, :stop], %{headers: headers}}
-      assert {"x-foo-response", "bar-response"} in headers
-    end
-
-    test "reports response status code", %{bypass: bypass, client: client} do
-      {test_name, _arity} = __ENV__.function
-      self = self()
-
-      :telemetry.attach(
-        to_string(test_name),
-        [:finch, :recv, :stop],
-        fn name, _, metadata, _ -> send(self, {:telemetry_event, name, metadata}) end,
-        nil
-      )
-
-      Bypass.expect(bypass, "GET", "/", fn conn -> Plug.Conn.send_resp(conn, 201, "OK") end)
-
-      request = Finch.build(:get, endpoint(bypass))
-      assert {:ok, %{status: 201}} = Finch.request(request, client)
-
-      assert_receive {:telemetry_event, [:finch, :recv, :stop], %{status: 201}}
-    end
-
-    test "reports reused connections", %{bypass: bypass, client: client} do
-      {test_name, _arity} = __ENV__.function
-      parent = self()
-      ref = make_ref()
-
-      handler = fn event, _measurements, meta, _config ->
-        case event do
-          [:finch, :connect, :start] ->
-            send(parent, {ref, :start})
-
-          [:finch, :connect, :stop] ->
-            send(parent, {ref, :stop})
-
-          [:finch, :reused_connection] ->
-            assert is_atom(meta.scheme)
-            assert is_binary(meta.host)
-            assert is_integer(meta.port)
-            send(parent, {ref, :reused})
-
-          _ ->
-            flunk("Unknown event")
-        end
-      end
-
-      :telemetry.attach_many(
-        to_string(test_name),
-        [
-          [:finch, :connect, :start],
-          [:finch, :connect, :stop],
-          [:finch, :reused_connection]
-        ],
-        handler,
-        nil
-      )
-
-      request = Finch.build(:get, endpoint(bypass))
-      assert {:ok, %{status: 200}} = Finch.request(request, client)
-      assert_receive {^ref, :start}
-      assert_receive {^ref, :stop}
-
-      assert {:ok, %{status: 200}} = Finch.request(request, client)
-      assert_receive {^ref, :reused}
-
-      :telemetry.detach(to_string(test_name))
-    end
-
-    test "reports conn_max_idle_time_exceeded", %{bypass: bypass, client: client} do
-      {test_name, _arity} = __ENV__.function
-      parent = self()
-      ref = make_ref()
-
-      handler = fn event, measurements, meta, _config ->
-        case event do
-          [:finch, :conn_max_idle_time_exceeded] ->
-            assert is_integer(measurements.idle_time)
-            assert is_atom(meta.scheme)
-            assert is_binary(meta.host)
-            assert is_integer(meta.port)
-            send(parent, {ref, :conn_max_idle_time_exceeded})
-
-          _ ->
-            flunk("Unknown event")
-        end
-      end
-
-      :telemetry.attach_many(
-        to_string(test_name),
-        [
-          [:finch, :conn_max_idle_time_exceeded]
-        ],
-        handler,
-        nil
-      )
-
-      request = Finch.build(:get, endpoint(bypass))
-      assert {:ok, %{status: 200}} = Finch.request(request, client)
-      Process.sleep(15)
-      assert {:ok, %{status: 200}} = Finch.request(request, client)
-      assert_receive {^ref, :conn_max_idle_time_exceeded}
-
-      :telemetry.detach(to_string(test_name))
-    end
-
-    test "reports max_idle_time_exceeded", %{bypass: bypass, client: client} do
-      {test_name, _arity} = __ENV__.function
-      parent = self()
-      ref = make_ref()
-
-      handler = fn event, measurements, meta, _config ->
-        case event do
-          [:finch, :max_idle_time_exceeded] ->
-            assert is_integer(measurements.idle_time)
-            assert is_atom(meta.scheme)
-            assert is_binary(meta.host)
-            assert is_integer(meta.port)
-            send(parent, {ref, :max_idle_time_exceeded})
-
-          _ ->
-            flunk("Unknown event")
-        end
-      end
-
-      :telemetry.attach_many(
-        to_string(test_name),
-        [
-          [:finch, :max_idle_time_exceeded]
-        ],
-        handler,
-        nil
-      )
-
-      request = Finch.build(:get, endpoint(bypass))
-      assert {:ok, %{status: 200}} = Finch.request(request, client)
-      Process.sleep(15)
-      assert {:ok, %{status: 200}} = Finch.request(request, client)
-      assert_receive {^ref, :max_idle_time_exceeded}
-
-      :telemetry.detach(to_string(test_name))
+      assert {:ok, _} = Finch.build(:get, endpoint(bypass)) |> Finch.request(finch_name)
     end
   end
 
   describe "stream/5" do
-    test "successful get request, with query string", %{bypass: bypass} do
-      start_supervised!({Finch, name: finch_name()})
+    test "successful get request, with query string", %{bypass: bypass, finch_name: finch_name} do
+      start_supervised!({Finch, name: finch_name})
       query_string = "query=value"
 
       Bypass.expect_once(bypass, "GET", "/", fn conn ->
@@ -1010,11 +564,12 @@ defmodule FinchTest do
 
       assert {:ok, {200, [_ | _], "OK"}} =
                Finch.build(:get, endpoint(bypass, "?" <> query_string))
-               |> Finch.stream(finch_name(), acc, fun)
+               |> Finch.stream(finch_name, acc, fun)
     end
 
-    test "successful post request, with query string and string request body", %{bypass: bypass} do
-      start_supervised!({Finch, name: finch_name()})
+    test "successful post request, with query string and string request body",
+         %{bypass: bypass, finch_name: finch_name} do
+      start_supervised!({Finch, name: finch_name})
       query_string = "query=value"
       req_headers = [{"content-type", "application/json"}]
       req_body = "{hello:\"world\"}"
@@ -1035,13 +590,15 @@ defmodule FinchTest do
 
       assert {:ok, {200, [_ | _], ^resp_body}} =
                Finch.build(:post, endpoint(bypass, "?" <> query_string), req_headers, req_body)
-               |> Finch.stream(finch_name(), acc, fun)
+               |> Finch.stream(finch_name, acc, fun)
     end
 
-    test "successful post request, with query string and streaming request body", %{
-      bypass: bypass
-    } do
-      start_supervised!({Finch, name: finch_name()})
+    test "successful post request, with query string and streaming request body",
+         %{
+           bypass: bypass,
+           finch_name: finch_name
+         } do
+      start_supervised!({Finch, name: finch_name})
       query_string = "query=value"
       req_headers = [{"content-type", "application/json"}]
       req_stream = Stream.map(1..10_000, fn _ -> "please" end)
@@ -1067,7 +624,7 @@ defmodule FinchTest do
                  req_headers,
                  {:stream, req_stream}
                )
-               |> Finch.stream(finch_name(), acc, fun)
+               |> Finch.stream(finch_name, acc, fun)
     end
   end
 
@@ -1075,19 +632,10 @@ defmodule FinchTest do
     Registry.lookup(name, shp)
   end
 
-  defp endpoint(%{port: port}, path \\ "/"), do: "http://localhost:#{port}#{path}"
-
   defp shp(%{port: port}), do: {:http, "localhost", port}
   defp shp({scheme, {:local, unix_socket}}), do: {scheme, {:local, unix_socket}, 0}
 
   defp expect_any(bypass) do
     Bypass.expect(bypass, fn conn -> Plug.Conn.send_resp(conn, 200, "OK") end)
-  end
-
-  # returns the name of the current test as an atom which can be used as the Finch name
-  # to enable async tests with many different instances of Finch running
-  defp finch_name do
-    {name, _} = __ENV__.function
-    name
   end
 end
